issue-193 - Move torch loader out of efficientnet code
